requirements.txt
openai>=1.0.0
pandas>=1.5.0
pydantic>=2.0.0
faiss-cpu>=1.7.0  # for vector store placeholder
python-dateutil>=2.8.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
pandas>=1.5.0
pydantic>=2.0.0
faiss-cpu>=1.7.0  # for vector store placeholder
python-dateutil>=2.8.0
python-dotenv>=1.0.0